Terminé comentarios rust
