+++ Compilador/lib/parser/comentario.rs
@@ -11,7 +11,7 @@ use nom::{
 ///
 /// # Parametros
 ///
-/// * `input`- Input a parsear
+/// * `input` - Input a parsear
 ///
 /// # Ejemplo
 ///
