+++ Compilador/lib/parser/reglas_expresion/exp.rs
@@ -41,7 +41,7 @@ fn checar_lista_operadores() {
 ///
 /// # Parametros
 ///
-/// * `input`- Input a parsear
+/// * `input` - Input a parsear
 ///
 /// # Gramática
 ///
