+++ Compilador/lib/scanners/mod.rs
@@ -1,8 +1,7 @@
-//! Módulo que se encarga del analisis de los terminales del lenguaje.
+//! Módulo que se encarga del analisis léxico del lenguaje.
 
 pub mod operadores;
 pub mod tipos;
 pub mod ws;
 pub mod id;
-pub mod texto;
 pub mod constantes;
